added updateGatewayTest
